move sideboard to component
